wip: extract app errors
