+++ src/apps/main/preload.js
@@ -297,5 +297,4 @@ contextBridge.exposeInMainWorld('electron', {
     return ipcRenderer.invoke('APP:PREFERRED_LANGUAGE');
   },
   path,
-  ipcRenderer,
 });
