+++ src/apps/renderer/hooks/GeneralIssues.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { AppIssue } from '../../shared/types';
+import { AppIssue } from '../../../shared/issues/AppIssue';
 
 export default function useGeneralIssues() {
   const [generalIssues, setGeneralIssues] = useState<AppIssue[]>([]);
