+++ src/apps/renderer/messages/general-error.ts
@@ -1,18 +1,18 @@
-import { AppErrorName } from '../../../shared/issues/AppIssue';
+import { AppError } from '../../../shared/issues/AppIssue';
 
-type GeneralErrorMessages = Record<AppErrorName, string>;
+type AppErrorMessages = Record<AppError, string>;
 
-export const shortMessages: GeneralErrorMessages = {
+export const shortMessages: AppErrorMessages = {
   UNKNOWN_DEVICE_NAME: 'Could not retrieve your device’s name.',
 };
 
-export const longMessages: GeneralErrorMessages = {
+export const longMessages: AppErrorMessages = {
   UNKNOWN_DEVICE_NAME: 'Please restart the app.',
 };
 
 export const generalErrors: {
-  longMessages: GeneralErrorMessages;
-  shortMessages: GeneralErrorMessages;
+  longMessages: AppErrorMessages;
+  shortMessages: AppErrorMessages;
 } = {
   shortMessages,
   longMessages,
