+++ src/apps/shared/types.ts
@@ -1,16 +1,3 @@
-import { AppErrorName } from '../../shared/issues/AppIssue';
-
-export type AppIssue = {
-  action: 'GET_DEVICE_NAME_ERROR';
-  errorName: AppErrorName;
-  process: 'GENERAL';
-  errorDetails: {
-    name: string;
-    message: string;
-    stack: string;
-  };
-};
-
 type DriveOperationInProgress = {
   action: 'UPLOADING' | 'DOWNLOADING' | 'RENAMING' | 'DELETING';
   progress: number;
@@ -23,15 +10,3 @@ type DriveOperationCompleted = {
 };
 
 export type DriveInfo = DriveOperationInProgress | DriveOperationCompleted;
-
-type SyncActionName =
-  | 'renameInLocal'
-  | 'renameInRemote'
-  | 'pullFromLocal'
-  | 'pullFromRemote'
-  | 'deleteInLocal'
-  | 'deleteInRemote';
-
-type SyncAction = Array<[string, string]> | Array<string>;
-
-export type EnqueuedSyncActions = Partial<Record<SyncActionName, SyncAction>>;
