--- src/context/desktop/app/domain/AppIssue.ts
@@ -1,10 +1,0 @@
-export type AppIssue = {
-  action: 'GET_DEVICE_NAME_ERROR';
-  errorName: 'UNKNOWN_DEVICE_NAME';
-  process: 'GENERAL';
-  errorDetails: {
-    name: string;
-    messages: string;
-    stack: string;
-  };
-};
